Arduino-python-piano-with-mqtt/python_code/requirements.txt
pysynth~=2.4
pyserial~=3.4
paho-mqtt~=1.5.0